Add tests for Minitest & Cucumber reporting hooks
This file is going to be need to be a bit more fleshed out, if we want
similar coverage for Minitest and Cucumber as we do for RSpec. The file
is still rather focused around RSpec.

I'm not sure why the after block is here; it prevents adding the other
examples, so I removed it.